chore: extract pod creation functionality into separate function Signed-off-by: Anton Misskii <amisskii@redhat.com>
